parse future date
