+++ pkg/prompts/prompts.go
@@ -6,18 +6,13 @@ import (
 	"errors"
 	"fmt"
 	"math/big"
-	"net/mail"
 	"net/url"
-	"os"
 	"strconv"
 	"time"
 
 	"github.com/ava-labs/avalanche-cli/pkg/constants"
 	"github.com/ava-labs/avalanche-cli/pkg/models"
-	"github.com/ava-labs/avalanche-cli/pkg/ux"
 	"github.com/ava-labs/avalanchego/ids"
-	avago_constants "github.com/ava-labs/avalanchego/utils/constants"
-	"github.com/ava-labs/avalanchego/utils/formatting/address"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/manifoldco/promptui"
 	"golang.org/x/mod/semver"
@@ -54,6 +49,7 @@ type Prompter interface {
 	CaptureWeight(promptStr string) (uint64, error)
 	CaptureUint64(promptStr string) (uint64, error)
 	CapturePChainAddress(promptStr string, network models.Network) (string, error)
+	CaptureFutureDate(promptStr string, minDate time.Time) (time.Time, error)
 }
 
 type realPrompter struct{}
@@ -61,97 +57,6 @@ type realPrompter struct{}
 // NewProcessChecker creates a new process checker which can respond if the server is running
 func NewPrompter() Prompter {
 	return &realPrompter{}
-}
-
-func validateEmail(input string) error {
-	_, err := mail.ParseAddress(input)
-	return err
-}
-
-func validatePositiveBigInt(input string) error {
-	n := new(big.Int)
-	n, ok := n.SetString(input, 10)
-	if !ok {
-		return errors.New("invalid number")
-	}
-	if n.Cmp(big.NewInt(0)) == -1 {
-		return errors.New("invalid number")
-	}
-	return nil
-}
-
-func validateStakingDuration(input string) error {
-	d, err := time.ParseDuration(input)
-	if err != nil {
-		return err
-	}
-	if d > constants.MaxStakeDuration {
-		return fmt.Errorf("exceeds maximum staking duration of %s", ux.FormatDuration(constants.MaxStakeDuration))
-	}
-	if d < constants.MinStakeDuration {
-		return fmt.Errorf("below the minimum staking duration of %s", ux.FormatDuration(constants.MinStakeDuration))
-	}
-	return nil
-}
-
-func validateTime(input string) error {
-	t, err := time.Parse(constants.TimeParseLayout, input)
-	if err != nil {
-		return err
-	}
-	if t.Before(time.Now().Add(constants.StakingStartLeadTime)) {
-		return fmt.Errorf("time should be at least start from now + %s", constants.StakingStartLeadTime)
-	}
-	return err
-}
-
-func validateNodeID(input string) error {
-	_, err := ids.NodeIDFromString(input)
-	return err
-}
-
-func validateAddress(input string) error {
-	if !common.IsHexAddress(input) {
-		return errors.New("invalid address")
-	}
-	return nil
-}
-
-func validateExistingFilepath(input string) error {
-	if fileInfo, err := os.Stat(input); err == nil && !fileInfo.IsDir() {
-		return nil
-	}
-	return errors.New("file doesn't exist")
-}
-
-func validateWeight(input string) error {
-	val, err := strconv.ParseUint(input, 10, 64)
-	if err != nil {
-		return err
-	}
-	if val < constants.MinStakeWeight || val > constants.MaxStakeWeight {
-		return errors.New("the weight must be an integer between 1 and 100")
-	}
-	return nil
-}
-
-func validateBiggerThanZero(input string) error {
-	val, err := strconv.ParseUint(input, 10, 64)
-	if err != nil {
-		return err
-	}
-	if val == 0 {
-		return errors.New("the value must be bigger than zero")
-	}
-	return nil
-}
-
-func validateURL(input string) error {
-	_, err := url.ParseRequestURI(input)
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 // CaptureListDecision runs a for loop and continuously asks the
@@ -314,68 +219,6 @@ func (*realPrompter) CapturePositiveBigInt(promptStr string) (*big.Int, error) {
 		return nil, errors.New("SetString: error")
 	}
 	return amountInt, nil
-}
-
-func validatePChainAddress(input string) (string, error) {
-	chainID, hrp, _, err := address.Parse(input)
-	if err != nil {
-		return "", err
-	}
-
-	if chainID != "P" {
-		return "", errors.New("this is not a PChain address")
-	}
-	return hrp, nil
-}
-
-func validatePChainFujiAddress(input string) error {
-	hrp, err := validatePChainAddress(input)
-	if err != nil {
-		return err
-	}
-	if hrp != avago_constants.FujiHRP {
-		return errors.New("this is not a fuji address")
-	}
-	return nil
-}
-
-func validatePChainMainAddress(input string) error {
-	hrp, err := validatePChainAddress(input)
-	if err != nil {
-		return err
-	}
-	if hrp != avago_constants.MainnetHRP {
-		return errors.New("this is not a mainnet address")
-	}
-	return nil
-}
-
-func validatePChainLocalAddress(input string) error {
-	hrp, err := validatePChainAddress(input)
-	if err != nil {
-		return err
-	}
-	// ANR uses the `custom` HRP for local networks,
-	// but the `local` HRP also exists...
-	if hrp != avago_constants.LocalHRP && hrp != avago_constants.FallbackHRP {
-		return errors.New("this is not a local nor custom address")
-	}
-	return nil
-}
-
-func getPChainValidationFunc(network models.Network) func(string) error {
-	switch network {
-	case models.Fuji:
-		return validatePChainFujiAddress
-	case models.Mainnet:
-		return validatePChainMainAddress
-	case models.Local:
-		return validatePChainLocalAddress
-	default:
-		return func(string) error {
-			return errors.New("unsupported network")
-		}
-	}
 }
 
 func (*realPrompter) CapturePChainAddress(promptStr string, network models.Network) (string, error) {
@@ -545,6 +388,35 @@ func (*realPrompter) CaptureIndex(promptStr string, options []any) (int, error) 
 		return 0, err
 	}
 	return listIndex, nil
+}
+
+// CaptureFutureDate requires from the user a date input which is in the future.
+// If `minDate` is not empty, the minimum time in the future from the provided date is required
+// Otherwise, time from time.Now() is chosen.
+func (*realPrompter) CaptureFutureDate(promptStr string, minDate time.Time) (time.Time, error) {
+	prompt := promptui.Prompt{
+		Label: promptStr,
+		Validate: func(s string) error {
+			t, err := time.Parse(constants.TimeParseLayout, s)
+			if err != nil {
+				return err
+			}
+			if minDate == (time.Time{}) {
+				minDate = time.Now()
+			}
+			if t.Before(minDate.UTC()) {
+				return fmt.Errorf("the provided date is before %s UTC", minDate.Format(constants.TimeParseLayout))
+			}
+			return nil
+		},
+	}
+
+	timestampStr, err := prompt.Run()
+	if err != nil {
+		return time.Time{}, err
+	}
+
+	return time.Parse(constants.TimeParseLayout, timestampStr)
 }
 
 func contains[T comparable](list []T, element T) bool {
